dockerLab/requirements.txt
Django
south
djangorestframework
markdown
django-filter
